core.common: move Json, datetime_aware, datetime_naive, is_namedtuple, asdict to my.core.types
