+++ my/core/time.py
@@ -3,7 +3,7 @@ from typing import Sequence, Dict
 
 import pytz
 
-from .common import datetime_aware, datetime_naive
+from .types import datetime_aware, datetime_naive
 
 
 def user_forced() -> Sequence[str]:
