+++ my/time/tz/main.py
@@ -1,8 +1,10 @@
 '''
 Timezone data provider, used to localize timezone-unaware timestamps for other modules
 '''
+
 from datetime import datetime
-from my.core.common import datetime_aware
+
+from my.core import datetime_aware
 
 # todo hmm, kwargs isn't mypy friendly.. but specifying types would require duplicating default args. uhoh
 def localize(dt: datetime, **kwargs) -> datetime_aware:
